@@ -16,7 +16,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 
-version: 1.0.5-tv+180
+version: 1.1.10-tv+189
 
 environment:
   sdk: '>=2.17.0 <3.0.0'
@@ -31,16 +31,21 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
 
   # FIREBASE PACKAGES
-  firebase_core: ^2.4.0
-  firebase_messaging: ^14.2.0
-  firebase_analytics: ^10.0.8
-  firebase_crashlytics: ^3.0.8
+  firebase_core: ^2.1.1
+  firebase_messaging: ^14.0.3
+  firebase_analytics: ^10.0.3
+  firebase_crashlytics: ^3.0.3
 
   # STORAGE
   shared_preferences: ^2.0.13
+  hive: ^2.2.3
+  hive_flutter: ^1.1.0
+
+
 
   # UI
   cupertino_icons: ^1.0.4
+  font_awesome_flutter: ^10.3.0
   flutter_spinkit: ^5.1.0
   page_transition: ^2.0.5
   #modal_progress_hud: ^0.1.3
@@ -48,6 +53,11 @@ dependencies:
 
   # CONNECTIVITY
   http: ^0.13.4
+  dio: ^4.0.6
+  dio_cache_interceptor: ^3.3.1
+  dio_cache_interceptor_hive_store: ^3.2.0
+  cached_network_image: ^3.1.0
+
   share: ^2.0.4
   launch_review: ^3.0.1
   url_launcher: ^6.1.0
@@ -56,7 +66,7 @@ dependencies:
   uni_links: ^0.5.1
 
   # TOOLS
-  google_fonts: ^3.0.1
+  google_fonts: ^4.0.3
   package_info: ^2.0.2
   android_intent: ^2.0.2
   global_configuration: ^2.0.0-nullsafety.1
@@ -67,19 +77,38 @@ dependencies:
 
   # STATE MANAGEMENT
   provider: ^6.0.2
+  flutter_riverpod: ^2.1.3
   path_provider: ^2.0.9
+
+  animated_text_kit: ^4.2.2
+  auto_size_text: ^3.0.0
+  flutter_animate: ^3.0.0
+  marquee: ^2.2.3
+  text_scroll: ^0.1.0
+
+  easy_debounce: ^2.0.2+1
 
   # MAP
   #location: ^4.3.0
   geolocator: ^9.0.2
-#  webview_flutter: ^3.0.0
+  hijri: ^3.0.0
+  arabic_numbers: ^1.0.0
+  persian_number_utility: ^1.1.3
+
+  #  webview_flutter: ^3.0.0
 
   rive_splash_screen: ^0.1.0
-  lottie: ^2.1.0
-  flutter_svg: ^1.0.3
+  lottie: ^2.2.0
+  flutter_svg: ^2.0.0+1
   introduction_screen: ^3.0.2
   dots_indicator: ^2.1.0
   sizer: ^2.0.15
+
+  timeago_flutter: ^1.2.0
+
+  #  media
+  youtube_player_flutter: ^8.1.2
+  audiofileplayer: ^2.1.1
 
 dev_dependencies:
   flutter_test:
@@ -90,6 +119,8 @@ dev_dependencies:
 
 # The following section is specific to Flutter.
 flutter:
+
+  generate: true
 
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
@@ -109,12 +140,25 @@ flutter:
     - assets/animations/rive/
     - assets/animations/lottie/
     - assets/img/logo/
+    - assets/icon/
 
 
   fonts:
     - family: MawaqitIcons
       fonts:
         - asset: assets/fonts/MawaqitIcons.ttf
-
-flutter_intl:
-  enabled: true
+    - family: Hafs
+      fonts:
+        - asset: assets/fonts/original-hafs.ttf
+    - family: kufi
+      fonts:
+        - asset: assets/fonts/kufi-normal.ttf
+    - family: arial
+      fonts:
+        - asset: assets/fonts/ARIAL.TTF
+    - family: helvetica
+      fonts:
+        - asset: assets/fonts/Helvetica.ttf
+    - family: WeatherIcons
+      fonts:
+        - asset: assets/fonts/weathericons-regular-webfont.ttf
