@@ -6,7 +6,7 @@
 // ignore_for_file: lines_longer_than_80_chars
 
 import 'package:connectivity_for_web/connectivity_for_web.dart';
-//import 'package:firebase_analytics_web/firebase_analytics_web.dart';
+import 'package:firebase_analytics_web/firebase_analytics_web.dart';
 import 'package:firebase_core_web/firebase_core_web.dart';
 import 'package:firebase_messaging_web/firebase_messaging_web.dart';
 import 'package:geolocator_web/geolocator_web.dart';
@@ -20,7 +20,7 @@ import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 // ignore: public_member_api_docs
 void registerPlugins(Registrar registrar) {
   ConnectivityPlugin.registerWith(registrar);
-  //FirebaseAnalyticsWeb.registerWith(registrar);
+  FirebaseAnalyticsWeb.registerWith(registrar);
   FirebaseCoreWeb.registerWith(registrar);
   FirebaseMessagingWeb.registerWith(registrar);
   GeolocatorPlugin.registerWith(registrar);
