@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:global_configuration/global_configuration.dart';
-import 'package:mawaqit/generated/l10n.dart';
+import 'package:mawaqit/i18n/l10n.dart';
 import 'package:mawaqit/src/data/config.dart';
 import 'package:mawaqit/src/helpers/AnalyticsWrapper.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -9,7 +9,9 @@ class AppLanguage extends ChangeNotifier {
   Locale _appLocale = Locale('en', '');
 
   String languageName(String languageCode) =>
-      Config.language.firstWhere((element) => element['value'] == languageCode)['subtitle'] ?? '';
+      Config.language.firstWhere(
+          (element) => element['value'] == languageCode)['subtitle'] ??
+      '';
 
   String get currentLanguageName => languageName(_appLocale.languageCode);
 
@@ -18,7 +20,8 @@ class AppLanguage extends ChangeNotifier {
   Future<void> fetchLocale() async {
     var prefs = await SharedPreferences.getInstance();
     if (prefs.getString('language_code') == null) {
-      _appLocale = Locale('${GlobalConfiguration().getValue('defaultLanguage')}', '');
+      _appLocale =
+          Locale('${GlobalConfiguration().getValue('defaultLanguage')}', '');
       notifyListeners();
       return null;
     }
@@ -38,11 +41,15 @@ class AppLanguage extends ChangeNotifier {
       mosqueId: mosqueId,
     );
 
-    if (S.delegate.supportedLocales.indexOf(type) != -1) {
+    if (S.supportedLocales.indexOf(type) != -1) {
       _appLocale = type;
       await prefs.setString('language_code', type.languageCode);
       await prefs.setString('countryCode', type.countryCode ?? '');
     }
     notifyListeners();
+  }
+
+  bool isArabic() {
+    return _appLocale.languageCode == "ar" || _appLocale.languageCode == "ur";
   }
 }
