@@ -1,22 +1,22 @@
 import 'dart:async';
-import 'dart:io';
 import 'dart:isolate';
 
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:global_configuration/global_configuration.dart';
-import 'package:mawaqit/generated/l10n.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart' show ProviderScope;
+import 'package:mawaqit/i18n/l10n.dart';
 import 'package:mawaqit/i18n/AppLanguage.dart';
 import 'package:mawaqit/src/enum/connectivity_status.dart';
 import 'package:mawaqit/src/helpers/AnalyticsWrapper.dart';
 import 'package:mawaqit/src/helpers/AppRouter.dart';
 import 'package:mawaqit/src/helpers/ConnectivityService.dart';
-import 'package:mawaqit/src/helpers/HttpOverrides.dart';
+import 'package:mawaqit/src/helpers/HiveLocalDatabase.dart';
 import 'package:mawaqit/src/pages/SplashScreen.dart';
+import 'package:mawaqit/src/services/audio_manager.dart';
+import 'package:mawaqit/src/services/developer_manager.dart';
 import 'package:mawaqit/src/services/mosque_manager.dart';
 import 'package:mawaqit/src/services/settings_manager.dart';
 import 'package:mawaqit/src/services/theme_manager.dart';
@@ -27,8 +27,6 @@ Future<void> main() async {
   runZonedGuarded<Future<void>>(() async {
     WidgetsFlutterBinding.ensureInitialized();
     // Sizer
-
-    await GlobalConfiguration().loadFromAsset("configuration");
 
     await Firebase.initializeApp();
 
@@ -41,24 +39,7 @@ Future<void> main() async {
       );
     }).sendPort);
 
-    if (kDebugMode) {
-      await FirebaseCrashlytics.instance.setCrashlyticsCollectionEnabled(false);
-    } else {
-      await FirebaseCrashlytics.instance.setCrashlyticsCollectionEnabled(true);
-    }
-
-    HttpOverrides.global = MyHttpOverrides();
-    FocusManager.instance.highlightStrategy = FocusHighlightStrategy.alwaysTraditional;
-
-    // hide status bar
-    SystemChrome.setEnabledSystemUIMode(SystemUiMode.immersive);
-    SystemChrome.setSystemUIChangeCallback((systemOverlaysAreVisible) async {
-      if (systemOverlaysAreVisible) return;
-      await Future.delayed(Duration(seconds: 3));
-      SystemChrome.restoreSystemUIOverlays();
-    });
-
-    return runApp(MyApp());
+    return runApp(ProviderScope(child: MyApp()));
   }, (error, stack) => FirebaseCrashlytics.instance.recordError(error, stack));
 }
 
@@ -71,13 +52,20 @@ class MyApp extends StatelessWidget {
         ChangeNotifierProvider(create: (context) => AppLanguage()),
         ChangeNotifierProvider(create: (context) => MosqueManager()),
         ChangeNotifierProvider(create: (context) => SettingsManager()),
+        ChangeNotifierProvider(create: (context) => AudioManager()),
+        ChangeNotifierProvider(create: (context) => DeveloperManager()),
+        ChangeNotifierProvider(create: (context) => HiveManager()),
       ],
       child: Consumer<AppLanguage>(builder: (context, model, child) {
         return Sizer(builder: (context, orientation, size) {
           return StreamProvider(
             initialData: ConnectivityStatus.Offline,
-            create: (context) => ConnectivityService().connectionStatusController.stream.map((event) {
-              if (event == ConnectivityStatus.Wifi || event == ConnectivityStatus.Cellular) {
+            create: (context) => ConnectivityService()
+                .connectionStatusController
+                .stream
+                .map((event) {
+              if (event == ConnectivityStatus.Wifi ||
+                  event == ConnectivityStatus.Cellular) {
                 //todo check actual internet
               }
 
@@ -85,10 +73,17 @@ class MyApp extends StatelessWidget {
             }),
             child: Consumer<ThemeNotifier>(
               builder: (context, theme, _) => Shortcuts(
-                shortcuts: {SingleActivator(LogicalKeyboardKey.select): ActivateIntent()},
+                shortcuts: {
+                  SingleActivator(LogicalKeyboardKey.select): ActivateIntent()
+                },
                 child: MaterialApp(
                   themeMode: theme.mode,
-                  // themeMode: ThemeMode.dark,
+                  localeResolutionCallback: (locale, supportedLocales) {
+                    if (locale?.languageCode.toLowerCase() == 'ba')
+                      return Locale('en');
+
+                    return locale;
+                  },
                   theme: theme.lightTheme,
                   darkTheme: theme.darkTheme,
                   locale: model.appLocal,
@@ -100,7 +95,7 @@ class MyApp extends StatelessWidget {
                     GlobalMaterialLocalizations.delegate,
                     GlobalWidgetsLocalizations.delegate,
                   ],
-                  supportedLocales: S.delegate.supportedLocales,
+                  supportedLocales: S.supportedLocales,
                   debugShowCheckedModeBanner: false,
                   home: Splash(),
                 ),
