@@ -1,3 +1,4 @@
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart' hide TextDirection;
 import 'package:mawaqit/src/helpers/StringUtils.dart';
@@ -38,17 +39,24 @@ class _DrawerListTitle extends State<DrawerListTitle> {
       onFocusChange: (i) => setState(() => isFocused = i),
       child: ListTile(
         autofocus: widget.autoFocus,
-        tileColor: isFocused ? Theme.of(context).focusColor : Colors.transparent,
-        textColor: isFocused ? Colors.white : Theme.of(context).textTheme.bodyMedium?.color,
+        tileColor:
+            isFocused ? Theme.of(context).focusColor : Colors.transparent,
+        textColor: isFocused
+            ? Colors.white
+            : Theme.of(context).textTheme.bodyMedium?.color,
         title: Text(
-          widget.autoTranslate ? Intl.message(widget.text ?? '', name: widget.text?.toCamelCase) : widget.text ?? '',
+          widget.autoTranslate
+              ? Intl.message(widget.text ?? '', name: widget.text?.toCamelCase)
+              : widget.text ?? '',
           overflow: TextOverflow.ellipsis,
           style: TextStyle(fontSize: 15.0),
         ),
         leading: _renderIcon(
           widget.iconUrl,
           widget.icon,
-          isFocused ? Colors.white : Theme.of(context).textTheme.bodyMedium?.color,
+          isFocused
+              ? Colors.white
+              : Theme.of(context).textTheme.bodyMedium?.color,
         ),
         trailing: widget.onTap == null
             ? null
@@ -56,7 +64,9 @@ class _DrawerListTitle extends State<DrawerListTitle> {
                 Directionality.of(context) == TextDirection.ltr
                     ? Icons.keyboard_arrow_right
                     : Icons.keyboard_arrow_left,
-                color: isFocused ? Colors.white : Theme.of(context).textTheme.bodyMedium?.color,
+                color: isFocused
+                    ? Colors.white
+                    : Theme.of(context).textTheme.bodyMedium?.color,
               ),
         onTap: widget.onTap,
       ),
@@ -65,8 +75,8 @@ class _DrawerListTitle extends State<DrawerListTitle> {
 
   Widget _renderIcon(String? iconUrl, IconData icon, Color? color) {
     return iconUrl != null && iconUrl != ""
-        ? Image.network(
-            iconUrl,
+        ? CachedNetworkImage(
+            imageUrl: iconUrl,
             width: 20,
             height: 20,
             color: widget.forceThemeColor ? color : null,
