@@ -12,12 +12,12 @@ class SharedPref {
     return json.decode(value);
   }
 
-  save(String key, value) async {
+  Future<void> save(String key, value) async {
     final prefs = await SharedPreferences.getInstance();
     prefs.setString(key, json.encode(value));
   }
 
-  remove(String key) async {
+  Future<void> remove(String key) async {
     final prefs = await SharedPreferences.getInstance();
     prefs.remove(key);
   }
