@@ -1,18 +1,31 @@
 import 'dart:async';
+import 'dart:io';
 
+import 'package:dio/dio.dart';
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:mawaqit/generated/l10n.dart';
+import 'package:flutter/services.dart';
+import 'package:global_configuration/global_configuration.dart';
+import 'package:mawaqit/i18n/l10n.dart';
 import 'package:mawaqit/i18n/AppLanguage.dart';
 import 'package:mawaqit/src/helpers/AppRouter.dart';
+import 'package:mawaqit/src/helpers/HiveLocalDatabase.dart';
+import 'package:mawaqit/src/helpers/HttpOverrides.dart';
+import 'package:mawaqit/src/helpers/RelativeSizes.dart';
 import 'package:mawaqit/src/helpers/SharedPref.dart';
 import 'package:mawaqit/src/models/settings.dart';
-import 'package:mawaqit/src/pages/HomeScreen.dart';
+import 'package:mawaqit/src/pages/ErrorScreen.dart';
+import 'package:mawaqit/src/pages/home/OfflineHomeScreen.dart';
 import 'package:mawaqit/src/pages/onBoarding/OnBoardingScreen.dart';
 import 'package:mawaqit/src/services/mosque_manager.dart';
 import 'package:mawaqit/src/services/settings_manager.dart';
+import 'package:mawaqit/src/widgets/InfoWidget.dart';
 import 'package:provider/provider.dart';
 import 'package:rive_splash_screen/rive_splash_screen.dart';
 import 'package:wakelock/wakelock.dart';
+
+enum ErrorState { mosqueNotFound, noInternet, mosqueDataError }
 
 class Splash extends StatefulWidget {
   @override
@@ -21,7 +34,36 @@ class Splash extends StatefulWidget {
 
 class _SplashScreen extends State<Splash> {
   SharedPref sharedPref = SharedPref();
-  bool applicationProblem = false;
+  ErrorState? error;
+
+  // bool applicationProblem = false;
+
+  Future<void> initApplicationUI() async {
+    await GlobalConfiguration().loadFromAsset("configuration");
+    Wakelock.enable().catchError((e) {});
+
+    initHive();
+
+    SystemChrome.setPreferredOrientations([
+      DeviceOrientation.landscapeRight,
+      DeviceOrientation.landscapeLeft,
+    ]);
+
+    await FirebaseCrashlytics.instance
+        .setCrashlyticsCollectionEnabled(!kDebugMode);
+
+    HttpOverrides.global = MyHttpOverrides();
+    FocusManager.instance.highlightStrategy =
+        FocusHighlightStrategy.alwaysTraditional;
+
+    // hide status bar
+    SystemChrome.setEnabledSystemUIMode(SystemUiMode.immersive);
+    SystemChrome.setSystemUIChangeCallback((systemOverlaysAreVisible) async {
+      if (systemOverlaysAreVisible) return;
+      await Future.delayed(Duration(seconds: 3));
+      SystemChrome.restoreSystemUIOverlays();
+    });
+  }
 
   Future<Settings> _initSettings() async {
     await context.read<AppLanguage>().fetchLocale();
@@ -41,25 +83,69 @@ class _SplashScreen extends State<Splash> {
   /// navigates to first screen
   void _navigateToHome() async {
     try {
-      Wakelock.enable().catchError((e) {});
-
+      await initApplicationUI();
       var settings = await _initSettings();
-
       var goBoarding = await loadBoarding();
       var mosqueManager = context.read<MosqueManager>();
-      bool hasNoMosque = mosqueManager.mosqueId == null && mosqueManager.mosqueSlug == null;
+      bool hasNoMosque = mosqueManager.mosqueUUID == null;
 
       if (hasNoMosque || goBoarding && settings.boarding == "1") {
-        AppRouter.pushReplacement(OnBoardingScreen(settings));
+        AppRouter.pushReplacement(OnBoardingScreen());
       } else {
-        AppRouter.pushReplacement(HomeScreen(settings));
+        AppRouter.pushReplacement(OfflineHomeScreen());
+      }
+    } on DioError catch (e) {
+      if (e.response == null) {
+        print('no internet connection');
+        print(e.requestOptions.uri);
+        setState(() => error = ErrorState.noInternet);
+
+        // mosque not found
+      } else {
+        setState(() => error = ErrorState.mosqueNotFound);
+        // e.response!.data;
       }
     } catch (e) {
-      setState(() => applicationProblem = true);
+      setState(() => error = ErrorState.mosqueDataError);
+      rethrow;
     }
   }
 
+  /// reset the app
+  void _changeMosque() {
+    AppRouter.pushReplacement(OnBoardingScreen());
+  }
+
   Widget build(BuildContext context) {
+    RelativeSizes.instance.size = MediaQuery.of(context).size;
+
+    switch (error) {
+      case ErrorState.mosqueNotFound:
+        return ErrorScreen(
+          title: S.of(context).reset,
+          description: S.of(context).mosqueNotFoundMessage,
+          image: 'assets/img/icon_exit.png',
+          onTryAgain: _changeMosque,
+          tryAgainText: S.of(context).changeMosque,
+        );
+      case ErrorState.noInternet:
+        return ErrorScreen(
+          title: S.of(context).noInternet,
+          description: S.of(context).noInternetMessage,
+          image: 'assets/img/wifi.png',
+          onTryAgain: _navigateToHome,
+        );
+      case ErrorState.mosqueDataError:
+        return ErrorScreen(
+          title: S.of(context).error,
+          description: S.of(context).mosqueErrorMessage,
+          image: 'assets/img/icon_exit.png',
+          onTryAgain: _navigateToHome,
+        );
+      case null:
+        break;
+    }
+
     return Scaffold(
       body: Container(
         color: Colors.black,
@@ -70,32 +156,28 @@ class _SplashScreen extends State<Splash> {
               'assets/backgrounds/splash_screen_5.png',
               fit: BoxFit.cover,
             ),
-            Container(
-              width: double.infinity,
-              child: SplashScreen.callback(
-                isLoading: false,
-                onSuccess: (e) => _navigateToHome(),
-                onError: (error, stacktrace) {},
-                name: 'assets/animations/rive/mawaqit_logo_animation1.riv',
-                fit: BoxFit.cover,
-                startAnimation: 'idle',
-                loopAnimation: 'loading_light',
-                endAnimation: 'loaded_text_light',
-                width: 100,
-                height: 100,
-              ),
-            ),
-            if (applicationProblem)
-              Align(
-                alignment: Alignment(0, .3),
-                child: Text(
-                  S.of(context).backendError,
-                  textAlign: TextAlign.center,
-                  style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-                        color: Colors.redAccent.withOpacity(.6),
-                      ),
+            RepaintBoundary(
+              child: Container(
+                width: double.infinity,
+                child: SplashScreen.callback(
+                  isLoading: false,
+                  onSuccess: (e) => _navigateToHome(),
+                  onError: (error, stacktrace) {},
+                  name: 'assets/animations/rive/mawaqit_logo_animation1.riv',
+                  fit: BoxFit.cover,
+                  startAnimation: 'idle',
+                  loopAnimation: 'loading_light',
+                  endAnimation: 'loaded_text_light',
+                  width: 100,
+                  height: 100,
                 ),
               ),
+            ),
+            Container(
+              padding: EdgeInsets.all(10),
+              child: Opacity(child: VersionWidget(), opacity: 1),
+              alignment: Alignment.bottomCenter,
+            ),
           ],
         ),
       ),
