@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mawaqit/generated/l10n.dart';
+import 'package:mawaqit/i18n/l10n.dart';
 
 class OnBoardingMawaqitAboutWidget extends StatelessWidget {
   const OnBoardingMawaqitAboutWidget({Key? key}) : super(key: key);
